refactor index for shared/api snapmomentAPI
