+++ src/shared/api/index.ts
@@ -1,4 +1,4 @@
 export * from './auth/authApi';
 export * from './common/model/api.types';
-export * from './common/snapmomentAPI';
+// export * from './common/snapmomentAPI';
 // export * from './common/snapmomentBaseQuery';
